requirements.txt
numpy~=1.24.4
taichi~=1.6.0
pyyaml~=6.0.1